Created search bar and basic CSS styling. Moved navbar and searchbar into filter directory
