+++ src/components/posts/Posts.css
@@ -63,24 +63,3 @@
     color: #555; /* Slightly lighter color for user name */
   }
   
-  .filter-bar {
-    display: flex;
-    justify-content: center; /* Center the filter bar horizontally */
-    align-items: center; /* Center items vertically */
-    gap: 20px; /* Space items evenly */
-    border: 2px solid purple; /* Add a purple border */
-    padding: 10px; /* Add some padding for better spacing */
-  }
-  
-  .filter-dropdown {
-    padding: 5px; /* Add some padding to dropdowns */
-    font-size: 16px; /* Adjust font size */
-    border-radius: 4px;
-  }
-  
-  .filter-checkbox {
-    display: flex;
-    align-items: center; /* Align checkbox and label vertically */
-    font-size: 16px; /* Adjust font size */
-  }
-
